Enhance docs tool (#309)
* Updated sidebar menu items to use expand icons

* Updated weight to use double instead of int

* Added configurable port option

* Added .nojekyll file generation and option to turn off

* Added sitemap.xml generation

* Switched to Apache commons io for majority of file operations for stability

* Added publish method using JGit

+++ scala-debugger-docs/src/main/scala/org/scaladebugger/docs/structures/Metadata.scala
@@ -27,7 +27,7 @@ import scala.util.Try
 case class Metadata(
   layout: String,
   usingDefaultLayout: Boolean,
-  weight: Int,
+  weight: Double,
   render: Boolean,
   title: Option[String],
   link: Option[String],
@@ -56,7 +56,7 @@ object Metadata {
     val redirect = data.get("redirect").flatMap(_.headOption)
     val layout = data.get("layout").flatMap(_.headOption)
     val weight = data.get("weight").flatMap(_.headOption)
-      .flatMap(w => Try(w.toInt).toOption)
+      .flatMap(w => Try(w.toDouble).toOption)
     val render = data.get("render").flatMap(_.headOption)
       .flatMap(r => Try(r.toBoolean).toOption)
     val fake = data.get("fake").flatMap(_.headOption)
